Undo update to migration

+++ db/migrate/20210808171558_create_bulk_discounts.rb
@@ -1,7 +1,7 @@
 class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
-      t.integer :discount
+      t.decimal :discount
       t.integer :quant_threshold
       t.references :merchant, foreign_key: true
